add support for deployment and inialization

+++ resources/initialize_db.py
@@ -1,3 +1,5 @@
+import os
+
 from app import Base
 from sqlalchemy import create_engine
 
@@ -7,6 +9,9 @@ from resources.dao.distributions_dao import Distribution
 from resources.dao.requests_dao import Request
 
 if __name__ == '__main__':
-    engine = create_engine('postgres://guest:guest@192.168.99.100:5432/guest')
-    Base.metadata.drop_all(engine)
+    if 'POSTGRES_URI' not in os.environ:
+        raise RuntimeError('Create POSTGRES_URI environment variable (format: postgres://user:password@host.port/db).')
+    engine = create_engine(os.environ['POSTGRES_URI'])
+    if 'CLEAN_FIRST' in os.environ and os.environ['CLEAN_FIRST'].lower() == 'true':
+        Base.metadata.drop_all(engine)
     Base.metadata.create_all(engine)
